💾 fix: Altera tipo de coluna de tinyint para smallint em 'num_doors' para resolver o conflito de tipo de dado com programa Java

+++ src/main/resources/db/migration/V2__create_car_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS `car` (
     `id` VARCHAR(36) PRIMARY KEY NOT NULL,
-    `num_doors` TINYINT NOT NULL,
+    `num_doors` SMALLINT NOT NULL,
     `engine_capacity` DECIMAL(2,1) NOT NULL,
     `car_type` VARCHAR(10) NOT NULL,
     FOREIGN KEY (`id`) REFERENCES `vehicle`(`id`)
